backend_challenge/requirements.txt
beautifulsoup4==4.12.2
Django==4.2.2
djangorestframework==3.14.0
django-crontab==0.7.1
drf-spectacular==0.26.3
Requests==2.31.0
